requirements.txt
PyPDF2
python-docx
wikipedia
beautifulsoup4
requests
lxml